+++ src/main/java/chess/WebUIChessApplication.java
@@ -1,5 +1,12 @@
 package chess;
 
+import chess.db.dao.BlackPieceDAO;
+import chess.db.dao.ChessBoardStateDAO;
+import chess.db.dao.WhitePieceDAO;
+import chess.service.ChessGameCalculatorScore;
+import chess.service.createInitialBoard;
+import chess.service.ChessGameMove;
+import chess.service.ChessGameSetting;
 import chess.webController.*;
 
 import static spark.Spark.staticFileLocation;
@@ -9,9 +16,13 @@ public class WebUIChessApplication {
         staticFileLocation("templates");
 
         ChessMainPageUrlController.run();
-        ChessFirstStartUrlController.run();
-        ChessContinueStartUrlController.run();
-        ChessMoveUrlController.run();
-        ChessStatusUrlController.run();
+
+        ChessFirstStartUrlController.run(new createInitialBoard());
+
+        ChessContinueStartUrlController.run(new ChessGameSetting());
+
+        ChessMoveUrlController.run(new ChessGameMove());
+
+        ChessStatusUrlController.run(new ChessGameCalculatorScore());
     }
 }
