+++ src/main/java/chess/webController/ChessStatusUrlController.java
@@ -1,15 +1,14 @@
 package chess.webController;
 
-import chess.service.ChessGameService;
+import chess.service.ChessGameCalculatorScore;
 
 import static spark.Spark.post;
 
 public class ChessStatusUrlController {
-    public static void run() {
-        ChessGameService chessGameService = new ChessGameService();
-
+    public static void run(ChessGameCalculatorScore chessGameCalculatorScore) {
         post("/status", (req, res) -> {
-            res.body(String.format("%s점수: %.1f", chessGameService.getColor(), chessGameService.calculateScore()));
+            res.body(String.format("%s점수: %.1f", chessGameCalculatorScore.getColor(),
+                    chessGameCalculatorScore.calculateScore()));
             return res.body();
         });
     }
